Add crude windows compatability, - Rename HttpHandler to linHttpHandler - Add new winHttpHandler for windows - Export duplicate functions from win/lin-HttpHandler into IHttpHandler
